Develop
* rename package details in control file

* control file details customised for karmalb

* script not needed as function is in utils

* add scripts to create a basic package

* add mini-httpd to install list

* add lb package and new dependencies

* use a cache for downloaded packages

* bump misc packages (for mini-httpd)

* missed two packages

* need to make pkg cache directory

* upgrade media debs to ensure consistency

* remove pkg debug code and remove regex

* make output prettier

* gnore distro pkgcache directory

* add flag to keep tmp files (for debugging)

* move path of load balancer files under /opt

* remove tmp directory if it was left behind

* move path of load balancer files under /opt

* makefile to build packages

* hack a mini-httpd build

* Makefile for top level

* remove workdir (unless -k is specified)

* cleanup directory on exit

* tidy up leftover build files

* ignore more debs and working files

* ln args wrong way around + use -e

* need checkinstall package

* fix patch offsets

* add abort on error flag

* make tar less verbose

* add more patches to fix warnings and build errors

* remove ISO just before making a new one

* add link to mini_httpd daemon in /sbin

* fix up networking if we started with DHCP

+++ src/karmalb-pkg/util/getzensource_from_iso.sh
@@ -1,10 +1,27 @@
 #/!bin/sh -e
-TMPDIR=tmp
+PKGNAME=karmalb
+PKGVERS=1.0a2
+PKGMAIN="KarmaLB Developers <dev@karmalb.org.uk>"
+OLDPREFIX=usr/local
+NEWPREFIX=opt
+OLDPATH=${OLDPREFIX}/zenloadbalancer
+NEWPATH=${NEWPREFIX}/klb
+#
 ZENPKG=zenloadbalancer_3.10.deb
 ZENISO=zenloadbalancer-distro_3.10.iso
+#
+TMPDIR=tmp
 FILELIST=../filelist
 FILESDIR=../files
 CONTROLDIR=../control
+PKGARCH=`dpkg-architecture -qDEB_BUILD_ARCH`
+#
+KEEPFILES=0
+
+if [ "x$1" = "x-k" ]; then
+	KEEPFILES=1
+	shift
+fi
 
 getyn()
 {
@@ -36,8 +53,14 @@ if [ ! -f $ZENPKG ]; then
 fi
 
 # extract files from package 
+rm -rf $TMPDIR
 mkdir $TMPDIR
 dpkg -x $ZENPKG $TMPDIR
+
+# move lb directory
+mkdir -p ${TMPDIR}/${NEWPREFIX}
+mv ${TMPDIR}/${OLDPATH} ${TMPDIR}/${NEWPATH}
+( cd $TMPDIR; rmdir -p $OLDPREFIX )
 
 # build filelist file
 ( cd $TMPDIR; find * -print ) | sort | \
@@ -62,16 +85,27 @@ mkdir $FILESDIR
 cat $FILELIST | (
 	while read T F X; do
 		case $T in
-			\#*) echo "Skipping binary $F";;
-			l) echo "Skipping link $F -> $X";;
-			d) mkdir $FILESDIR/$F;;
-			f) cp -p $TMPDIR/$F $FILESDIR/$F;;
+			\#*)	echo "Skipping binary $F";;
+			l)	echo "Skipping link $F -> $X";;
+			d)	mkdir $FILESDIR/$F;;
+			f)	cp -p $TMPDIR/$F $FILESDIR/$F
+				if grep -q /$OLDPATH $FILESDIR/$F; then
+					sed -i "s@/$OLDPATH@/$NEWPATH@g" $FILESDIR/$F
+					echo "Modified $F"
+				fi
+				;;
 		esac
 	done
 )
-rm -rf $TMPDIR
+test $KEEPFILES -eq 1 || rm -rf $TMPDIR
 
 # extract control files from package
 rm -rf $CONTROLDIR
 mkdir $CONTROLDIR
 dpkg -e $ZENPKG $CONTROLDIR
+# fix up control file
+sed -i -e "s/^Package:.*/Package: $PKGNAME/" \
+	-e "s/^Version:.*/Version: $PKGVERS/" \
+	-e "s/^Maintainer:.*/Maintainer: $PKGMAIN/" \
+	-e "s/^Architecture:.*/Architecture: $PKGARCH/" \
+	$CONTROLDIR/control
